fix: move profile logic to Profile class
